+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Updated custom data schema
 - Data checks now performed after optimisation
 - Difficulty related function now uses wrapper instead of concrete object
+- Renamed save option attributes from `optimise` to `optimize`
 
 ### Fixed